third_party/blink/web_tests/external/wpt/editing/run/italic_2001-last-expected.txt

This is a testharness.js-based test.
Found 14 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","false"],["italic",""]] "<span style=font-style:italic>fo[o</span><span style=font-style:oblique>b]ar</span>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "<span style=font-style:italic>fo[o</span><span style=font-style:oblique>b]ar</span>" queryCommandState("italic") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["italic",""]] "<span style=font-style:italic>fo[o</span><span style=font-style:oblique>b]ar</span>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "<span style=font-style:oblique>fo[o</span><span style=font-style:italic>b]ar</span>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["italic",""]] "<span style=font-style:oblique>fo[o</span><span style=font-style:italic>b]ar</span>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "<span style=font-style:oblique>fo[o</span><span style=font-style:italic>b]ar</span>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "<span style=font-style:oblique>fo[o</span><span style=font-style:italic>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-style:oblique\\">fo</span>ob<i>ar</i>" but got "<span style=\\"font-style:oblique\\">fo</span>ob<span style=\\"font-style:italic\\">ar</span>"
[FAIL] [["stylewithcss","false"],["italic",""]] "<span style=font-style:oblique>fo[o</span><span style=font-style:italic>b]ar</span>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "<span style=font-style:oblique>fo[o</span><span style=font-style:italic>b]ar</span>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "<i>fo[o</i><address>b]ar</address>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-style:italic\\">fo</span>o<address><span style=\\"font-style:normal\\">b</span>ar</address>" but got "<i>fo</i>o<address><span style=\\"font-style:normal\\">b</span>ar</address>"
[FAIL] [["stylewithcss","true"],["italic",""]] "<i>fo[o</i><address>b]ar</address>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "<i>fo[o</i><address>b]ar</address>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "<i>fo[o</i><address>b]ar</address>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "<i>fo[o</i><address>b]ar</address>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
Harness: the test ran to completion.
